dconf modification FINALLY works
It took a long time to figure out, but it turns out that upon login,
Arch Linux only executes `~/.profile` if `~/.bash_profile` does not
exist. This meant that this file that I had been modifying was doing
nothing! At last, I know this and the script works very well now. I'm
very glad I don't have to deal with this any more.